+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Config option to disable spawning with a material book
 ### Fixed
 - Material list in Material Book now clips text too long for the page (such as missing translations) [#882]
+- Material book will now show icons for materials with only part substitutes (such as blaze rod, etc.) [#882]
 
 ## [1.21.1-4.1.0.2] - 2026-02-16
 ### Fixed